Remove circular depencency between maps_legacy and region_map and tile_map plugins (#88265)
* Remove circular depencency between maps_legacy and region_map and tile_map plugins

* update imports in maps plugin

* tslint

# Conflicts:
#	src/plugins/region_map/config.ts
#	src/plugins/tile_map/config.ts
